Add ds/nonstd.h with some useful type traits
