+++ src/ds/serializer.h
@@ -59,16 +59,6 @@ namespace serializer
       }
     }
 
-    // C++20
-    template <class T>
-    struct remove_cvref
-    {
-      typedef std::remove_cv_t<std::remove_reference_t<T>> type;
-    };
-
-    template <class T>
-    using remove_cvref_t = typename remove_cvref<T>::type;
-
     template <typename Tup>
     struct TupMatcher
     {
